+++ R/expr__expr.R
@@ -12,6 +12,7 @@
 #'  * Functions
 #'  * List
 #'  * Meta
+#'  * Name
 #'  * String
 #'  * Struct
 #'  * Temporal
@@ -701,7 +702,7 @@ construct_ProtoExprArray = function(...) {
 #' # map a,b,c,d sequentially
 #' pl$LazyFrame(a = 1, b = 2, c = 3, d = 4)$select(
 #'   pl$all()$map(\(s) {
-#'     Sys.sleep(.5)
+#'     Sys.sleep(.1)
 #'     s * 2
 #'   })
 #' )$collect() |> system.time()
@@ -712,7 +713,7 @@ construct_ProtoExprArray = function(...) {
 #' pl$options$rpool_cap
 #' pl$LazyFrame(a = 1, b = 2, c = 3, d = 4)$select(
 #'   pl$all()$map(\(s) {
-#'     Sys.sleep(.5)
+#'     Sys.sleep(.1)
 #'     s * 2
 #'   }, in_background = TRUE)
 #' )$collect() |> system.time()
@@ -721,7 +722,7 @@ construct_ProtoExprArray = function(...) {
 #' pl$options$rpool_cap
 #' pl$LazyFrame(a = 1, b = 2, c = 3, d = 4)$select(
 #'   pl$all()$map(\(s) {
-#'     Sys.sleep(.5)
+#'     Sys.sleep(.1)
 #'     s * 2
 #'   }, in_background = TRUE)
 #' )$collect() |> system.time()
@@ -792,8 +793,8 @@ Expr_map = function(f, output_type = NULL, agg_list = FALSE, in_background = FAL
 #' # apply over groups - normal usage
 #' # s is a series of all values for one column within group, here Species
 #' e_all = pl$all() # perform groupby agg on all columns otherwise e.g. pl$col("Sepal.Length")
-#' e_sum = e_all$apply(\(s)  sum(s$to_r()))$suffix("_sum")
-#' e_head = e_all$apply(\(s) head(s$to_r(), 2))$suffix("_head")
+#' e_sum = e_all$apply(\(s)  sum(s$to_r()))$name$suffix("_sum")
+#' e_head = e_all$apply(\(s) head(s$to_r(), 2))$name$suffix("_head")
 #' pl$DataFrame(iris)$group_by("Species")$agg(e_sum, e_head)
 #'
 #'
@@ -804,10 +805,12 @@ Expr_map = function(f, output_type = NULL, agg_list = FALSE, in_background = FAL
 #' e_all = pl$col(pl$dtypes$Float64)
 #' e_add10 = e_all$apply(\(x)  {
 #'   x + 10
-#' })$suffix("_sum")
+#' })$name$suffix("_sum")
 #' # quite silly index into alphabet(letters) by ceil of float value
 #' # must set return_type as not the same as input
-#' e_letter = e_all$apply(\(x) letters[ceiling(x)], return_type = pl$dtypes$Utf8)$suffix("_letter")
+#' e_letter = e_all$apply(\(x) {
+#'   letters[ceiling(x)]
+#' }, return_type = pl$dtypes$Utf8)$name$suffix("_letter")
 #' pl$DataFrame(iris)$select(e_add10, e_letter)
 #'
 #'
@@ -923,48 +926,6 @@ Expr_lit = function(x) {
   # use .call reduces eval from 22us to 15us, not a bottle-next anyways
   .Call(wrap__Expr__lit, x) |>
     unwrap("in $lit()")
-}
-
-#' Add a suffix to a column name
-#' @keywords Expr
-#'
-#' @param suffix Suffix to be added to column name(s)
-#' @rdname Expr_suffix
-#' @return Expr
-#' @aliases suffix
-#' @seealso
-#' [`$prefix()`][Expr_prefix] to add a prefix
-#' @examples
-#' dat = pl$DataFrame(mtcars)
-#'
-#' dat$select(
-#'   pl$col("mpg"),
-#'   pl$col("mpg")$suffix("_foo"),
-#'   pl$col("cyl", "drat")$suffix("_bar")
-#' )
-Expr_suffix = function(suffix) {
-  .pr$Expr$suffix(self, suffix)
-}
-
-#' Add a prefix to a column name
-#' @keywords Expr
-#'
-#' @param prefix Prefix to be added to column name(s)
-#' @rdname Expr_prefix
-#' @return Expr
-#' @aliases prefix
-#' @seealso
-#' [`$suffix()`][Expr_suffix] to add a suffix
-#' @examples
-#' dat = pl$DataFrame(mtcars)
-#'
-#' dat$select(
-#'   pl$col("mpg"),
-#'   pl$col("mpg")$prefix("foo_"),
-#'   pl$col("cyl", "drat")$prefix("bar_")
-#' )
-Expr_prefix = function(prefix) {
-  .pr$Expr$prefix(self, prefix)
 }
 
 #' polars reverse
@@ -1167,54 +1128,6 @@ Expr_exclude = function(columns) {
     or_else = pstop(err = paste0("this type is not supported for Expr_exclude: ", columns))
   )
 }
-
-
-# TODO contribute pypolars keep_name example does not showcase an example where the name changes
-#' Keep the original root name of the expression.
-#'
-#' @keywords Expr
-#' @return Expr
-#' @aliases keep_name
-#' @name Expr_keep_name
-#' @docType NULL
-#' @format NULL
-#' @format NULL
-#' @examples
-#' pl$DataFrame(list(alice = 1:3))$select(pl$col("alice")$alias("bob")$keep_name())
-Expr_keep_name = "use_extendr_wrapper"
-
-
-
-# TODO contribute polars, map_alias unwrap user function errors instead of passing them back
-#' Map alias of expression with an R function
-#'
-#' @param fun an R function which takes a string as input and return a string
-#'
-#' @description Rename the output of an expression by mapping a function over the root name.
-#' @keywords Expr
-#' @return Expr
-#' @aliases map_alias
-#' @name Expr_map_alias
-#' @examples
-#' pl$DataFrame(list(alice = 1:3))$select(
-#'   pl$col("alice")$alias("joe_is_not_root")$map_alias(\(x) paste0(x, "_and_bob"))
-#' )
-Expr_map_alias = function(fun) {
-  if (
-    !polars_optenv$no_messages &&
-      !exists(".warn_map_alias", envir = runtime_state)
-  ) {
-    assign(".warn_map_alias", 1L, envir = runtime_state)
-    # it does not seem map alias is executed multi-threaded but rather immediately during building lazy query
-    # if ever crashing, any lazy method like select, filter, with_columns must use something like filter_with_r_func_support()
-    message("map_alias function is experimentally without some thread-safeguards, please report any crashes") # TODO resolve
-  }
-  if (!is.function(fun)) pstop(err = "alias_map fun must be a function")
-  if (length(formals(fun)) == 0) pstop(err = "alias_map fun must take at least one parameter")
-  .pr$Expr$map_alias(self, fun)
-}
-
-
 
 #' Are elements finite
 #' @description Returns a boolean output indicating which values are finite.
@@ -1609,7 +1522,8 @@ Expr_sort = function(descending = FALSE, nulls_last = FALSE) { # param reverse n
 #' ))$select(pl$col("a")$top_k(5))
 Expr_top_k = function(k) {
   if (!is.numeric(k) || k < 0) stopf("k must be numeric and positive, prefereably integerish")
-  .pr$Expr$top_k(self, k)
+  .pr$Expr$top_k(self, k) |>
+    unwrap("in $top_k():")
 }
 
 # TODO contribute polars, add arguments for Null/NaN/inf last/first, bottom_k unwraps k> len column
@@ -1630,7 +1544,8 @@ Expr_top_k = function(k) {
 #' ))$select(pl$col("a")$bottom_k(5))
 Expr_bottom_k = function(k) {
   if (!is.numeric(k) || k < 0) stopf("k must be numeric and positive, prefereably integerish")
-  .pr$Expr$bottom_k(self, k)
+  .pr$Expr$bottom_k(self, k) |>
+    unwrap("in $bottom_k():")
 }
 
 
@@ -4337,6 +4252,21 @@ Expr_meta = method_as_property(function() {
   expr_meta_make_sub_ns(self)
 })
 
+#' name: related methods
+#' @description
+#' Create an object namespace of all name related methods.
+#' See the individual method pages for full details
+#' @keywords Expr
+#' @return Expr
+#' @aliases name_ns
+#' @examples
+#'
+#' # missing
+#'
+Expr_name = method_as_property(function() {
+  expr_name_make_sub_ns(self)
+})
+
 #' cat: related methods
 #' @description
 #' Create an object namespace of all cat related methods.
@@ -4420,4 +4350,40 @@ Expr_lit_to_s = function() {
 #' )
 Expr_lit_to_df = function() {
   pl$select(self)
+}
+
+#' Find local minima
+#'
+#' A local minimum is the point that marks the transition between a decrease
+#' and an increase in a Series. The first and last values of the Series can never
+#' be a peak.
+#'
+#' @return Expr
+#' @seealso `$peak_max()`
+#'
+#' @examples
+#' df = pl$DataFrame(x = c(1, 2, 3, 2, 3, 4, 5, 2))
+#' df
+#'
+#' df$with_columns(peak_min = pl$col("x")$peak_min())
+Expr_peak_min = function() {
+  .pr$Expr$peak_min(self)
+}
+
+#' Find local maxima
+#'
+#' A local maximum is the point that marks the transition between an increase
+#' and a decrease in a Series. The first and last values of the Series can never
+#' be a peak.
+#'
+#' @return Expr
+#' @seealso `$peak_min()`
+#'
+#' @examples
+#' df = pl$DataFrame(x = c(1, 2, 3, 2, 3, 4, 5, 2))
+#' df
+#'
+#' df$with_columns(peak_max = pl$col("x")$peak_max())
+Expr_peak_max = function() {
+  .pr$Expr$peak_max(self)
 }