+++ R/lazyframe__group_by.R
@@ -34,7 +34,7 @@ print.LazyGroupBy = function(x, ...) {
 #' lgb$
 #'
 #'   agg(
-#'   pl$col("bar")$sum()$suffix("_sum"),
+#'   pl$col("bar")$sum()$name$suffix("_sum"),
 #'   pl$col("bar")$mean()$alias("bar_tail_sum")
 #' )
 LazyGroupBy_agg = agg = function(...) {