+++ R/options.R
@@ -52,6 +52,8 @@ polars_optreq$rpool_cap = list() # rust-side options already check args
 #' @docType NULL
 #'
 #' @details
+#' All args must be explicitly and fully named.
+#'
 #' `pl$options$rpool_active` indicates the number of R sessions already
 #' spawned in pool. `pl$options$rpool_cap` indicates the maximum number of new R
 #' sessions that can be spawned. Anytime a polars thread worker needs a background
@@ -100,8 +102,19 @@ pl$set_options = function(
   # modified in the first call)
   args_modified = names(as.list(sys.call()[-1]))
 
+  if (is.null(args_modified) || any(nchar(args_modified) == 0L)) {
+    Err_plain("all args must be named") |>
+      unwrap("in pl$set_options")
+  }
+
   for (i in seq_along(args_modified)) {
-    value = get(args_modified[i])
+    value = result(get(args_modified[i])) |>
+      map_err(\(rp_err) {
+        rp_err$
+          hint("arg-name does not match any defined args of `?set_options`")$
+          bad_arg(args_modified[i])
+      }) |>
+      unwrap("in pl$set_options")
 
     # each argument has its own input requirements
     validation = c()
@@ -180,29 +193,57 @@ subtimer_ms = function(cap_name = NULL, cap = 9999) {
 
 ### Other options implemented on rust side (likely due to thread safety)
 
-#' Toggle the global string cache
+#' Enable the global string cache
 #'
 #' Some functions (e.g joins) can be applied on Categorical series only allowed
-#' if using the global string cache is enabled. This function enables or disables
-#' the string_cache and override any contexts made by `pl$with_string_cache()`.
-#' In general, you should use `pl$with_string_cache()` instead.
+#' if using the global string cache is enabled. This function enables
+#' the string_cache. In general, you should use `pl$with_string_cache()` instead.
 #'
 #' @name pl_enable_string_cache
 #'
 #' @keywords options
-#' @param toggle Boolean. TRUE enable, FALSE disable.
-#' @return enable_string_cache: no return
+#' @return This doesn't return any value.
 #' @seealso
 #' [`pl$using_string_cache`][pl_using_string_cache]
+#' [`pl$disable_string_cache`][pl_disable_string_cache]
 #' [`pl$with_string_cache`][pl_with_string_cache]
 #' @examples
-#' pl$enable_string_cache(TRUE)
+#' pl$enable_string_cache()
 #' pl$using_string_cache()
-pl$enable_string_cache = function(toggle) {
-  enable_string_cache(toggle) |>
+#' pl$disable_string_cache()
+#' pl$using_string_cache()
+pl$enable_string_cache = function() {
+  enable_string_cache() |>
     unwrap("in pl$enable_string_cache()") |>
     invisible()
 }
+
+
+#' Disable the global string cache
+#'
+#' Some functions (e.g joins) can be applied on Categorical series only allowed
+#' if using the global string cache is enabled. This function disables
+#' the string_cache. In general, you should use `pl$with_string_cache()` instead.
+#'
+#' @name pl_disable_string_cache
+#'
+#' @keywords options
+#' @return This doesn't return any value.
+#' @seealso
+#' [`pl$using_string_cache`][pl_using_string_cache]
+#' [`pl$enable_string_cache`][pl_enable_string_cache]
+#' [`pl$with_string_cache`][pl_with_string_cache]
+#' @examples
+#' pl$enable_string_cache()
+#' pl$using_string_cache()
+#' pl$disable_string_cache()
+#' pl$using_string_cache()
+pl$disable_string_cache = function() {
+  disable_string_cache() |>
+    unwrap("in pl$disable_string_cache()") |>
+    invisible()
+}
+
 
 
 #' Check if the global string cache is enabled
@@ -217,9 +258,9 @@ pl$enable_string_cache = function(toggle) {
 #' [`pl$with_string_cache`][pl_with_string_cache]
 #' [`pl$enable_enable_cache`][pl_enable_string_cache]
 #' @examples
-#' pl$enable_string_cache(TRUE)
+#' pl$enable_string_cache()
 #' pl$using_string_cache()
-#' pl$enable_string_cache(FALSE)
+#' pl$disable_string_cache()
 #' pl$using_string_cache()
 pl$using_string_cache = function() {
   using_string_cache()
@@ -245,8 +286,8 @@ pl$using_string_cache = function() {
 #' })
 #' pl$concat(list(df1, df2))
 pl$with_string_cache = function(expr) {
-  increment_string_cache_counter(TRUE)
-  on.exit(increment_string_cache_counter(FALSE))
+  token = .pr$RPolarsStringCacheHolder$hold()
+  on.exit(token$release()) # if token was not release on exit, would release later on gc()
   eval(expr, envir = parent.frame())
 }
 
@@ -263,26 +304,28 @@ pl$with_string_cache = function(expr) {
 #' Multi-process communication has overhead because all data must be
 #' serialized/de-serialized and sent via buffers. Using multiple R sessions
 #' will likely only give a speed-up in a `low io - high cpu` scenario. Native
-#' polars query syntax runs in threads and have no overhead.
+#' polars query syntax runs in threads and have no overhead. Polars has as default
+#' double as many thread workers as cores. If any worker are queuing for or using R sessions,
+#' other workers can still continue any native polars parts as much as possible.
 #'
 #' @return
-#' `pl$get_global_rpool_cap()` returns a list with two elements `available`
-#' and `capacity`. `available` is the number of R sessions are already spawned
-#' in pool. `capacity` is the limit of new R sessions to spawn. Anytime a polars
+#' `pl$options$rpool_cap` returns the capacity ("limit") of co-running external R sessions /
+#' processes. `pl$options$rpool_active` is the number of R sessions are already spawned
+#' in the pool. `rpool_cap` is the limit of new R sessions to spawn. Anytime a polars
 #' thread worker needs a background R session specifically to run R code embedded
 #' in a query via `$map(..., in_background = TRUE)` or
 #' `$apply(..., in_background = TRUE)`, it will obtain any R session idling in
-#' rpool, or spawn a new R session (process) and add it to pool if `capacity`
+#' rpool, or spawn a new R session (process) if `capacity`
 #' is not already reached. If `capacity` is already reached, the thread worker
-#' will sleep until an R session is idling.
+#' will sleep and in a R job queue until an R session is idle.
 #'
 #' @keywords options
 #' @examples
-#' default = pl$get_global_rpool_cap()
-#' print(default)
-#' pl$set_global_rpool_cap(8)
-#' pl$get_global_rpool_cap()
-#' pl$set_global_rpool_cap(default$capacity)
+#' default = pl$options$rpool_cap |> print()
+#' pl$set_options(rpool_cap = 8)
+#' pl$options$rpool_cap
+#' pl$set_options(rpool_cap = default)
+#' pl$options$rpool_cap
 pl$get_global_rpool_cap = function() {
   warning(
     "in pl$get_global_rpool_cap(): Deprecated. Use pl$options$rpool_cap instead.",