+++ R/rbackground.R
@@ -65,7 +65,7 @@ print.RThreadHandle = function(x, ...) as.character(x) |> cat("\n")
 #' [`<Expr>$apply()`][Expr_apply]
 #' @examples
 #' prexpr = pl$col("mpg")$map(\(x) {
-#'   Sys.sleep(1.5)
+#'   Sys.sleep(.1)
 #'   x * 0.43
 #' }, in_background = TRUE)$alias("kml")
 #' handle = pl$LazyFrame(mtcars)$with_columns(prexpr)$collect_in_background()