+++ src/rust/src/lib.rs
@@ -25,6 +25,8 @@ pub mod rlib;
 pub mod rpolarserr;
 pub mod rstringcache;
 pub mod series;
+#[cfg(feature = "sql")]
+pub mod sql;
 pub mod utils;
 pub use serde_json;
 
@@ -39,6 +41,7 @@ static CONFIG: ThreadComStorage = InitCell::new();
 pub use crate::rbackground::RBGPOOL;
 
 // Macro to generate exports
+#[cfg(not(feature = "sql"))]
 extendr_module! {
     mod polars;
     use rlib;
@@ -48,6 +51,21 @@ extendr_module! {
     use rbackground;
     use lazy;
     use series;
+    use info;
+    use rstringcache;
+}
+
+#[cfg(feature = "sql")]
+extendr_module! {
+    mod polars;
+    use rlib;
+    use concat;
+    use rdataframe;
+    use rpolarserr;
+    use rbackground;
+    use lazy;
+    use series;
+    use sql;
     use info;
     use rstringcache;
 }