+++ tests/testthat/test-options.R
@@ -35,4 +35,20 @@ test_that("pl$options$ read-write", {
   # reset_options() works
   pl$reset_options()
   expect_identical(pl$options, old_options)
+
+  # all set_options args must be named
+  expect_identical(
+    pl$set_options(42) |> get_err_ctx("Plain"),
+    "all args must be named"
+  )
+  expect_identical(
+    pl$set_options(rpool_cap = 42, 42) |> get_err_ctx("Plain"),
+    "all args must be named"
+  )
+
+  # incomplete/misspelled name not allowed
+  expect_identical(
+    pl$set_options(rpo = 42) |> get_err_ctx("Hint"),
+    "arg-name does not match any defined args of `?set_options`"
+  )
 })